requirements.txt
boto3==1.26.0
botocore==1.29.0
aws-cdk-lib==2.50.0
constructs==10.1.0
pdfplumber==0.7.6
openpyxl==3.0.10
pandas==1.5.3
streamlit==1.12.0
statsmodels==0.13.5
moto==4.1.10
localstack-client==2.0
